Moved module on upper level, renamed directory
