--- 01_search/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "editor.fontSize": 16
-}
